omp works fine thread sanitizer gives false positives

+++ include/particle.hpp
@@ -31,7 +31,7 @@ public:
                             std::mt19937& rnd_gen) const;
     void MakeGasCollision(const double distance,
                           std::mt19937& rnd_gen);
-    int Trace(std::vector<std::unique_ptr<Surface>>& walls, const Background& gas,
+    size_t Trace(std::vector<std::unique_ptr<Surface>>& walls, const Background& gas,
                   std::mt19937& rnd_gen);
     Vec3 GetRandomVel(const Vec3& direction, std::mt19937& rnd_gen) const;
